initial working code for human detect

+++ stm32handC5_411CEU6/Src/auto_trip_50hz.c
@@ -21,8 +21,8 @@
  *                double res[2560]
  * Return Type  : void
  */
-void auto_trip_50hz(const double src[2560], double fs, double lamda, const
-                    double peaks[3], double res[2560])
+void auto_trip_50hz(const float src[2560], double fs, double lamda, const
+                    double peaks[3], float res[2560])
 {
   int i;
   int b_i;
